Removed JS code to set background image and added CSS styles
