+++ blocks/cards/cards.js
@@ -1,18 +1,13 @@
-import { createOptimizedPicture } from '../../scripts/aem.js';
-
 export default function decorate(block) {
-  /* change to ul, li */
-  const ul = document.createElement('ul');
   [...block.children].forEach((row) => {
-    const li = document.createElement('li');
-    while (row.firstElementChild) li.append(row.firstElementChild);
-    [...li.children].forEach((div) => {
-      if (div.children.length === 1 && div.querySelector('picture')) div.className = 'cards-card-image';
-      else div.className = 'cards-card-body';
-    });
-    ul.append(li);
+    row.classList.add('card');
+    const cardBackgroundImage = row.querySelectorAll('div')[0];
+    cardBackgroundImage.classList.add('card-background-image');
+
+    const cardHeader = row.querySelectorAll('div')[1];
+    cardHeader.classList.add('card-header');
+
+    const cardLinkIcon = row.querySelectorAll('div')[2];
+    cardLinkIcon.classList.add('card-link-icon');
   });
-  ul.querySelectorAll('picture > img').forEach((img) => img.closest('picture').replaceWith(createOptimizedPicture(img.src, img.alt, false, [{ width: '750' }])));
-  block.textContent = '';
-  block.append(ul);
 }
