--- blocks/landing-card/landing-card.css
@@ -1,139 +1,0 @@
-h5, p {
-  margin: 0;
-  padding: 0;
-}
-
-
-.landing-card  {
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  gap: .625rem;
-  margin: 0 auto;  
-}
-
-.industry-card {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  align-self: stretch;
-}
-
-.card {
-  display: flex;
-  width: 100%;
-  max-width: 400px;  
-  flex-direction: column;
-  justify-content: space-between;
-  box-sizing: border-box;
-  border: 1px solid #ddd;
-  border-radius: var(--border-tear-drop);
-  padding: 2rem 2.3rem;
-  background: linear-gradient(180deg, rgba(1 27 43 / 100%) 0%, rgba(0 19 30 / 0%) 100%);
-  transition: background-image 0.5s ease-in-out;
-  min-height: 400px;
-}
-
-.card:hover {
-  background-image: url('../../assets/card-background.svg'); 
-  transform: translateY(-5px);
-}
-
-.industry-card .card {
-  width: 19.4942rem;
-  min-height: 28.67rem;
-}
-
-.card-header {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5754rem;
-}
-
-.card-header > h5 {
-  color: var(--text-inverse);
-}
-
-.card-link-icon > p {
-  display: flex;
-  justify-content: flex-end;
-  cursor: pointer;
-
-} 
-
-.card-link-icon a.button:any-link,
-.card-link-icon > p > span {
-  all:unset
-}
-
-.icon-cardarrow {
-  display: contents;
-}
-
-.icon-cardarrow > img {
-  width: 2.4453rem;
-  height: 2.4453rem;
-}
-
-.card-header > p > picture > a > img {
-  width: 2.4933rem;
-  height: 2.4933rem;
-}
-
-@media (768px <= width <= 1024px) {
-  .landing-card {
-      flex-flow: row wrap;
-      justify-content: center;
-      gap: .625rem;
-  }
-
-  .card {
-      flex: 1;
-      max-width: 32%; /* Two cards per row */
-      padding: 1rem 1.3rem;
-      min-height: auto;
-  }
-
-  .industry-card .card {
-    flex: 1 1 calc(50% - 16px); /* 2 cards per row */
-    max-width: calc(50% - 16px);
-    min-height: 28.6695rem;
-  }
-}
-
-@media (width > 1024px){  
-  .card {
-    max-width: 25rem;
-    flex: 1 1 calc(25% - 16px);
-  }
-  
-  .landing-card .card {
-    width: 25.5rem;
-    height: 28.375rem;
-    padding: 2.5rem 3rem;
-  }
-
-  .industry-card .card {
-    width: 18rem;
-    height: 28rem;
-    padding: 2.5rem;
-    flex: 1 1 calc(25% - 16px);
-    max-width: calc(25% - 16px);
-  } 
-
-  .card-header {
-    gap:0.75rem;
-  }
-
-  .icon-cardarrow > img {
-    width: 3.1875rem;
-    height: 3.1875rem;
-  }
-
-  .card-header > p > picture > a > img {
-    width: 3.25rem;
-    height: 3.25rem;
-  }
-
-}
-
